requirements.txt
python-telegram-bot==21.4
fastapi
uvicorn
python-dotenv